+++ app/src/main/java/com/mbj/doeat/ui/screen/home/detail/detail_home/DetailScreen.kt
@@ -49,7 +49,7 @@ import com.mbj.doeat.ui.component.button.BackButton
 import com.mbj.doeat.ui.component.loading.LoadingView
 import com.mbj.doeat.ui.component.button.LongRectangleButtonWithParams
 import com.mbj.doeat.ui.component.party.HomeDetailPartyContent
-import com.mbj.doeat.ui.component.ReusableWebView
+import com.mbj.doeat.ui.component.webview.ReusableWebView
 import com.mbj.doeat.ui.component.toast.ToastMessage
 import com.mbj.doeat.ui.component.dialog.YesNoDialog
 import com.mbj.doeat.ui.component.party.NoPartiesAvailable
