+++ app/src/main/java/com/mbj/doeat/ui/screen/home/detail/detail_writer/PartyDetailWriterScreen.kt
@@ -23,7 +23,7 @@ import com.mbj.doeat.ui.component.button.BackButton
 import com.mbj.doeat.ui.component.loading.LoadingView
 import com.mbj.doeat.ui.component.button.LongRectangleButtonWithParams
 import com.mbj.doeat.ui.component.party.PartyDetailContent
-import com.mbj.doeat.ui.component.ReusableWebView
+import com.mbj.doeat.ui.component.webview.ReusableWebView
 import com.mbj.doeat.ui.component.dialog.YesNoDialog
 import com.mbj.doeat.ui.screen.home.detail.detail_writer.viewmodel.PartyDetailWriterViewModel
 import com.mbj.doeat.ui.theme.Color.Companion.Red500
